requirements.txt
python-telegram-bot==21.6
openpyxl==3.1.5
python-dotenv==1.0.1
openai>=1.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
yfinance>=0.2.0
